Hal 3.0 ap enable fixes 1 (#21)
* Always remove config before adding new one

It may happen that iw wifiX info command doesn't show relevant
data, but hostapd still uses the device. If we don't remove
the config then we may hit the 'device busy' error in hostpad.

Signed-off-by: Michal Kowalczyk <mkowalczyk@plume.com>

* Enable AP as the last action before applySettings()

AP enable will reload config. Because of that, all settings should
be set before it's called. For example, the SSID must be set,
otherwise config will fail because it requires SSID name and PSK
if WPA security is enabled.

Signed-off-by: Michal Kowalczyk <mkowalczyk@plume.com>