Improve ACL handling
Previously wifi_getApAclDeviceNum simply returned RETURN_ERR. Implement
that funtion properly now.

For:
- wifi_addApAclDevice
- wifi_delApAclDevice
- wifi_delApAclDevices

Modify the ACL file, and also call hostapd cli if the AP interface is up.

For:
- wifi_getApAclDevices

Return the list based on the ACL file if the interface is down, and
based on hostapd_cli if the interface is up.

That way the ACL state is always up to date, irrespective if modifications
are done when the AP interface is up or down.

Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>
Signed-off-by: Piotr Nakraszewicz <pnakraszewicz@plume.com>